remove MapView and react-native-maps
It's not currently used, and the dependency's presence causes the Apple
processor to get upset about not declaring NSLocationWhenInUseUsageDescription:
https://github.com/react-native-maps/react-native-maps/issues/3305

My original idea here was to show all regions-by-status on a world map.
But there's a further problem with the map library, at least on iOS: The
maximum zoom level of the map only shows 1 hemisphere of the globe,
making it misleading (eg cannot show IAD and SYD at the same time).

Revisit later!